# docs/swagger.yaml
basePath: /
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  request.SendMessageRequest:
    properties:
      operation:
        enum:
        - START
        - STOP
        type: string
    required:
    - operation
    type: object
  response.PaginatedResponse:
    properties:
      messages: {}
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
    type: object
info:
  contact: {}
  description: API for automatic message sending system
  title: Automatic Message Sending Service
  version: "1.0"
paths:
  /messages:
    get:
      consumes:
      - application/json
      description: Get a paginated list of sent messages
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "500":
          description: Internal server error response
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get sent messages
      tags:
      - messages
  /send:
    put:
      consumes:
      - application/json
      description: Start or stop the message sending process
      parameters:
      - description: Operation details
        in: body
        name: operation
        required: true
        schema:
          $ref: '#/definitions/request.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"operation": "START"} or {"operation": "STOP"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error response
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Send message operation
      tags:
      - messages
swagger: "2.0"